LLM_Server/requirements.txt
# LLM Server Requirements
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
accelerate>=0.20.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.0
langchain>=0.0.200
openai>=0.28.0
tiktoken>=0.5.0